Parse expressions!
Finally wrapped my head around the loop in the Pratt/TDOP parser. If we
parse a left prefix and find that it should be the left subtree of the
next token then we set the left-hand side to this just parsed result and
keep on parsing. We'll build up a left-biased tree this way.