Support function definitions and calls
Use a classic interpreter design with a separate call frame stack from
value stack. Locals are stored referenced to a stack base which is
pushed into call frame stack. On return the stack is restored to state
prior to call and result is pushed onto stack.

Changed semantics of instruction pointer, points to _next_ instruction
to execute when we're executing an instruction, this makes it much
easier to deal with absolute location jumps.